requirements.txt
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
greenlet==1.1.2
pydantic==1.9.1
Pygments==2.12.0
python-decouple==3.6
rich==12.5.1
shellingham==1.5.0
SQLAlchemy==1.4.40
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
typer==0.6.1
typing_extensions==4.3.0
